requirements.txt
pandas
sqlalchemy
psycopg2-binary
faker
pytest
jupyter
